added Pull Request statuses
