+++ app/models/notification.rb
@@ -3,8 +3,10 @@ class Notification < ApplicationRecord
 
   include Octobox::Notifications::InclusiveScope
   include Octobox::Notifications::ExclusiveScope
+  include Octobox::Notifications::SyncSubject
+  include Octobox::Notifications::SyncRepository
 
-  SUBJECTABLE_TYPES = ['Issue', 'PullRequest', 'Commit', 'Release'].freeze
+  SUBJECTABLE_TYPES = SUBJECT_TYPE_COMMIT_RELEASE + SUBJECT_TYPE_ISSUE_REQUEST.values
 
   if DatabaseConfig.is_postgres?
     include PgSearch
@@ -147,124 +149,5 @@ class Notification < ApplicationRecord
 
   def display_subject?
     @display_subject ||= subjectable? && (Octobox.fetch_subject? || github_app_installed?)
-  end
-
-  def update_subject(force = false)
-    return unless display_subject?
-    return if !force && subject != nil && updated_at - subject.updated_at < 2.seconds
-
-    UpdateSubjectWorker.perform_async_if_configured(self.id, force)
-  end
-
-  def update_subject_in_foreground(force = false)
-    return unless display_subject?
-    # skip syncing if the notification was updated around the same time as subject
-    return if !force && subject != nil && updated_at - subject.updated_at < 2.seconds
-
-    remote_subject = download_subject
-    return unless remote_subject.present?
-
-    if subject
-      case subject_type
-      when 'Issue', 'PullRequest'
-        subject.repository_full_name = repository_full_name
-        subject.assignees = ":#{Array(remote_subject.assignees.try(:map, &:login)).join(':')}:"
-        subject.state = remote_subject.merged_at.present? ? 'merged' : remote_subject.state
-        subject.save(touch: false) if subject.changed?
-      end
-    else
-      case subject_type
-      when 'Issue', 'PullRequest'
-        create_subject({
-          repository_full_name: repository_full_name,
-          github_id: remote_subject.id,
-          state: remote_subject.merged_at.present? ? 'merged' : remote_subject.state,
-          author: remote_subject.user.login,
-          html_url: remote_subject.html_url,
-          created_at: remote_subject.created_at,
-          updated_at: remote_subject.updated_at,
-          assignees: ":#{Array(remote_subject.assignees.try(:map, &:login)).join(':')}:",
-          locked: remote_subject.locked,
-        })
-      when 'Commit', 'Release'
-        create_subject({
-          repository_full_name: repository_full_name,
-          github_id: remote_subject.id,
-          author: remote_subject.author&.login,
-          html_url: remote_subject.html_url,
-          created_at: remote_subject.created_at,
-          updated_at: remote_subject.updated_at,
-          locked: remote_subject.locked
-        })
-      end
-    end
-
-    case subject_type
-    when 'Issue', 'PullRequest'
-      subject.update_labels(remote_subject.labels) if remote_subject.labels.present?
-    end
-  end
-
-  def update_repository(force = false)
-    return unless Octobox.config.subjects_enabled?
-    return if !force && repository != nil && updated_at - repository.updated_at < 2.seconds
-
-    UpdateRepositoryWorker.perform_async_if_configured(self.id, force)
-  end
-
-  def update_repository_in_foreground(force = false)
-    return unless Octobox.config.subjects_enabled?
-    return if !force && repository != nil && updated_at - repository.updated_at < 2.seconds
-
-    remote_repository = download_repository
-
-    if remote_repository.nil?
-      # if we can't access the repository, assume that it's private
-      remote_repository = OpenStruct.new({
-        full_name: repository_full_name,
-        private: true,
-        owner: {login: repository_owner_name}
-      })
-    end
-
-    if repository
-      repository.update_attributes({
-        full_name: remote_repository.full_name,
-        private: remote_repository.private,
-        owner: remote_repository.owner[:login],
-        github_id: remote_repository.id,
-        last_synced_at: Time.current
-      })
-    else
-      create_repository({
-        full_name: remote_repository.full_name,
-        private: remote_repository.private,
-        owner: remote_repository.owner[:login],
-        github_id: remote_repository.id,
-        last_synced_at: Time.current
-      })
-    end
-  end
-
-  private
-
-  def download_subject
-    user.subject_client.get(subject_url)
-
-  # If permissions changed and the user hasn't accepted, we get a 401
-  # We may receive a 403 Forbidden or a 403 Not Available
-  # We may be rate limited and get a 403 as well
-  # We may also get blocked by legal reasons (451)
-  # Regardless of the reason, any client error should be rescued and warned so we don't
-  # end up blocking other syncs
-  rescue Octokit::ClientError => e
-    Rails.logger.warn("\n\n\033[32m[#{Time.now}] WARNING -- #{e.message}\033[0m\n\n")
-    nil
-  end
-
-  def download_repository
-    user.github_client.repository(repository_full_name)
-  rescue Octokit::ClientError
-    nil
   end
 end
