+++ db/migrate/20181002064359_add_status_to_subjects.rb
@@ -1,0 +1,5 @@
+class AddStatusToSubjects < ActiveRecord::Migration[5.2]
+  def change
+    add_column :subjects, :status, :string
+  end
+end
